+++ src/pages/index.ts
@@ -1,13 +1,13 @@
 import { IncidentDetailsPage } from "./incident/IncidentDetails";
 import { IncidentListPage } from "./incident/IncidentListPage";
 import { LogsPage } from "./logs-page";
-import { TopologyPage } from "./topology-page";
 import { CanaryPage } from "./canary";
 import { ConfigListPage } from "./config/config-list";
 import { ConfigDetailsPage } from "./config/config-details";
 import { ConfigChangesPage } from "./config/config-changes";
 import { ConfigDetailsChangesPage } from "./config/config-details-changes";
 import { TimelinePage } from "./timeline-page";
+import { TopologyPage } from "./TopologyPage";
 
 export {
   LogsPage,
